Split balanced and unbalanced Merkle Tree tests into separate modules
